+++ src/lib.rs
@@ -21,6 +21,7 @@ pub mod bar;
 pub mod bind;
 pub mod buildpool;
 pub mod cgroup;
+pub mod config;
 pub mod context;
 pub mod ctl;
 pub mod deliver;